Move table config types to a new module
The `Internal` module is growing rather large, so I am planning on moving
functionality related to merge scheduling to a new module. For that, we first
need to move the config types to prevent cyclic dependencies.
