+++ lsm-tree.cabal
@@ -115,6 +115,7 @@ library
     Database.LSMTree.Internal.BloomFilter
     Database.LSMTree.Internal.ByteString
     Database.LSMTree.Internal.Chunk
+    Database.LSMTree.Internal.Config
     Database.LSMTree.Internal.CRC32C
     Database.LSMTree.Internal.Entry
     Database.LSMTree.Internal.IndexCompact
